- fixed compilation on macOS
error: non-constant-expression cannot be narrowed from type 'unsigned long' to 'int' in initializer list
error: unknown type name 'OpenGLFrameBuffer'

+++ src/hwrenderer/data/flatvertices.cpp
@@ -82,8 +82,8 @@ FFlatVertexBuffer::FFlatVertexBuffer(int width, int height)
 	mVertexBuffer->SetData(bytesize, nullptr, false);
 
 	static const FVertexBufferAttribute format[] = {
-		{ 0, VATTR_VERTEX, VFmt_Float3, myoffsetof(FFlatVertex, x) },
-		{ 0, VATTR_TEXCOORD, VFmt_Float2, myoffsetof(FFlatVertex, u) }
+		{ 0, VATTR_VERTEX, VFmt_Float3, (int)myoffsetof(FFlatVertex, x) },
+		{ 0, VATTR_TEXCOORD, VFmt_Float2, (int)myoffsetof(FFlatVertex, u) }
 	};
 	mVertexBuffer->SetFormat(1, 2, sizeof(FFlatVertex), format);
