+++ src/hwrenderer/scene/hw_skydome.cpp
@@ -88,9 +88,9 @@ FSkyVertexBuffer::FSkyVertexBuffer()
 	mVertexBuffer = screen->CreateVertexBuffer();
 
 	static const FVertexBufferAttribute format[] = {
-		{ 0, VATTR_VERTEX, VFmt_Float3, myoffsetof(FSkyVertex, x) },
-		{ 0, VATTR_TEXCOORD, VFmt_Float2, myoffsetof(FSkyVertex, u) },
-		{ 0, VATTR_COLOR, VFmt_Byte4, myoffsetof(FSkyVertex, color) }
+		{ 0, VATTR_VERTEX, VFmt_Float3, (int)myoffsetof(FSkyVertex, x) },
+		{ 0, VATTR_TEXCOORD, VFmt_Float2, (int)myoffsetof(FSkyVertex, u) },
+		{ 0, VATTR_COLOR, VFmt_Byte4, (int)myoffsetof(FSkyVertex, color) }
 	};
 	mVertexBuffer->SetFormat(1, 3, sizeof(FSkyVertex), format);
 	mVertexBuffer->SetData(mVertices.Size() * sizeof(FSkyVertex), &mVertices[0], true);
